+++ tests/testthat/test-maps.R
@@ -24,9 +24,10 @@ test_that("SpecimenMap R6 Subclass", {
 
   # Border Features ------------------------------------------------------------
   expect_type(vouchers$features, type = "closure")
-  voucher_features <- vouchers$features(sf_border = "black")
+  expect_error(vouchers$features(sf_border = "black", sf_states = list()))
 
   # County / State layer simple features
+  voucher_features <- vouchers$features(sf_border = "black")
   purrr::walk(
     .x = list(
       voucher_features[[1]][[1]],